Move AlphaRatio provider to sickrage.providers module
